Remove MockTransport, replace with cassette-based tests (#1808)
