+++ logfire/_internal/integrations/claude_agent_sdk.py
@@ -359,8 +359,8 @@ def instrument_claude_agent_sdk(logfire_instance: Logfire) -> AbstractContextMan
             span_data[INPUT_MESSAGES] = input_messages
         if hasattr(self, 'options') and self.options:  # pragma: no branch
             system_prompt = getattr(self.options, 'system_prompt', None)
-            if system_prompt:
-                text = system_prompt if isinstance(system_prompt, str) else str(system_prompt)
+            if system_prompt:  # pragma: no branch
+                text = str(system_prompt)
                 span_data[SYSTEM_INSTRUCTIONS] = [TextPart(type='text', content=text)]
 
         with logfire_claude.span('invoke_agent', **span_data) as root_span:
@@ -475,7 +475,7 @@ class _TurnTracker:
 
         # Per-turn token usage from AssistantMessage.usage
         usage = getattr(message, 'usage', None)
-        if usage:
+        if usage:  # pragma: no branch
             for key, value in _extract_usage(usage).items():
                 span_data[key] = value
 
@@ -485,26 +485,26 @@ class _TurnTracker:
 
         # Set error if the assistant message indicates an error
         error = getattr(message, 'error', None)
-        if error:
+        if error:  # pragma: no cover
             self._current_span.set_attribute('error.type', str(error))
 
         # Reset pending input — next turn will collect tool results
         self._pending_input = []
 
     def close(self) -> None:
-        if self._current_span is not None:
+        if self._current_span is not None:  # pragma: no branch
             self._current_span.__exit__(None, None, None)
             self._current_span = None
 
 
 def _record_result(span: LogfireSpan, msg: ResultMessage) -> None:
     """Record ResultMessage data onto the root span."""
-    if hasattr(msg, 'usage') and msg.usage:
+    if hasattr(msg, 'usage') and msg.usage:  # pragma: no branch
         usage = _extract_usage(msg.usage)
         for key, value in usage.items():
             span.set_attribute(key, value)
 
-    if hasattr(msg, 'total_cost_usd') and msg.total_cost_usd is not None:
+    if hasattr(msg, 'total_cost_usd') and msg.total_cost_usd is not None:  # pragma: no branch
         span.set_attribute('operation.cost', float(msg.total_cost_usd))
 
     session_id = getattr(msg, 'session_id', None)
@@ -516,5 +516,5 @@ def _record_result(span: LogfireSpan, msg: ResultMessage) -> None:
             span.set_attribute(attr, value)
 
     is_error = getattr(msg, 'is_error', None)
-    if is_error:
+    if is_error:  # pragma: no cover
         span.set_level('error')
